feat: Preload routes on home page
This commit preloads the routes for the "works", "contact", "random notes", and "playground" pages on the home page. This is done to improve the user experience by reducing the perceived loading time when navigating to these pages. The routes are preloaded using dynamic imports within a conditional check for interactive mode.